move the colibri grpc service out of pkg/cs
