+++ go/colibri/BUILD.bazel
@@ -23,8 +23,8 @@ go_library(
         "//go/lib/snet:go_default_library",
         "//go/lib/topology:go_default_library",
         "//go/pkg/app/launcher:go_default_library",
+        "//go/pkg/co/colibri/grpc:go_default_library",
         "//go/pkg/colibri/config:go_default_library",
-        "//go/pkg/cs/colibri/grpc:go_default_library",
         "//go/pkg/grpc:go_default_library",
         "//go/pkg/proto/colibri:go_default_library",
         "//go/pkg/storage:go_default_library",
