+++ go/colibri/main.go
@@ -37,8 +37,8 @@ import (
 	"github.com/scionproto/scion/go/lib/snet"
 	"github.com/scionproto/scion/go/lib/topology"
 	"github.com/scionproto/scion/go/pkg/app/launcher"
+	colgrpc "github.com/scionproto/scion/go/pkg/co/colibri/grpc"
 	"github.com/scionproto/scion/go/pkg/colibri/config"
-	colgrpc "github.com/scionproto/scion/go/pkg/cs/colibri/grpc"
 	libgrpc "github.com/scionproto/scion/go/pkg/grpc"
 	colpb "github.com/scionproto/scion/go/pkg/proto/colibri"
 	"github.com/scionproto/scion/go/pkg/storage"
