+++ go/pkg/co/colibri/grpc/BUILD.bazel
@@ -3,7 +3,7 @@ load("//lint:go.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["colibri_service.go"],
-    importpath = "github.com/scionproto/scion/go/pkg/cs/colibri/grpc",
+    importpath = "github.com/scionproto/scion/go/pkg/co/colibri/grpc",
     visibility = ["//visibility:public"],
     deps = [
         "//go/co/reservation:go_default_library",
